% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_hodler.R
\name{add_hodler}
\alias{add_hodler}
\title{Add hodler row to data event table}
\usage{
add_hodler(data, data_event)
}
\arguments{
\item{data}{stock data from load_data}

\item{data_event}{data with buy and sell events}
}
\description{
Add hodler row to data event table
}
