% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumsum_with_reset.R
\name{cumsum_with_reset}
\alias{cumsum_with_reset}
\title{Cumulative sum with reset}
\usage{
cumsum_with_reset(x)
}
\arguments{
\item{x}{binary vector}
}
\description{
`cumsum_with_reset` calculates the cumulative sum of x with 0 as the reset value
}
