% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_candlestick_legend}
\alias{get_candlestick_legend}
\title{Get candlestick legend}
\usage{
get_candlestick_legend(ticker = NULL)
}
\arguments{
\item{ticker}{string specifying the stock}
}
\value{
a string
}
\description{
Get candlestick legend
}
\details{
The function is a helper function for `plot.stock`
}
