% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_event_info}
\alias{get_event_info}
\title{Get event info}
\usage{
get_event_info(data_event, var_event, var_info)
}
\arguments{
\item{data_event}{data.frame with events obtained from `get_events`}

\item{var_event}{string specifying "buy_days" or "sell_days"}

\item{var_info}{string specifying which variable to select (should be "Date" or "Close")}
}
\description{
Get event info
}
\details{
This is a helper function for `plot.stock`
}
