% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_indicator}
\alias{get_indicator}
\title{Get indicator}
\usage{
get_indicator(data)
}
\arguments{
\item{data}{stock data}
}
\value{
a string specifying the selected indicator
}
\description{
Get indicator
}
\details{
The function is a helper function for `plot.stock`
}
