% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_fred.R
\name{get_percent_change}
\alias{get_percent_change}
\title{Get percent change}
\usage{
get_percent_change(x, from_start = FALSE, digits = 5)
}
\arguments{
\item{x}{numeric vector}

\item{from_start}{logical indicating if percent change should be calculated since start or day-by-day. Default is FALSE}

\item{digits}{number of digits using in rounding}
}
\value{
numeric vector with percent change
}
\description{
Get percent change
}
