% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_performance.R
\name{get_performance}
\alias{get_performance}
\title{Get performance}
\usage{
get_performance(data, strategy = "SMA", ...)
}
\arguments{
\item{data}{stock data}

\item{strategy}{string specifying indicator(s) to get performance based on.}

\item{...}{additional parameters passed to strategy definition using `buy_value` and `sell_value`}
}
\description{
Get performance
}
