% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_plot_info}
\alias{get_plot_info}
\title{Get plot info}
\usage{
get_plot_info(p, event, event_days, event_values)
}
\arguments{
\item{p}{plotly object}

\item{event}{string specifiying the event as "buy" or "sell"}

\item{event_days}{vector with dates for when the event is occurring}

\item{even_values}{vector with stock closing price at `event_days`}
}
\description{
Get plot info
}
\details{
This is a helper function for `plot.stock`
}
