% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_strategy.R
\name{get_strategy}
\alias{get_strategy}
\title{Get strategy}
\usage{
get_strategy(data, strategy, ...)
}
\arguments{
\item{data}{stock data}

\item{strategy}{string specifying indicator}

\item{...}{additional parameters passed to strategy definition using `buy_value` and `sell_value`}
}
\description{
Get strategy
}
