% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_strategy.R
\name{get_strategy_rsi}
\alias{get_strategy_rsi}
\title{Get strategy RSI}
\usage{
get_strategy_rsi(data, buy_value = 55, sell_value = 45)
}
\arguments{
\item{data}{stock data}

\item{buy_value}{value between 0 and 100 for which the strategy buys the stock. Default is 55}

\item{sell_value}{value between 0 and 100 for which the strategy sells the stock. Default is 45}
}
\description{
Get strategy RSI
}
