% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_fred.R
\name{load_fred}
\alias{load_fred}
\title{Loading data from Federal Reserve bank (FRED)}
\usage{
load_fred(symbol)
}
\arguments{
\item{symbol}{symbol to load}
}
\value{
data.frame with data from FRED
}
\description{
`load_fred` takes character symbol and loads data from Federal Reserve bank (FRED)
}
