% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_prediction}
\alias{plot_prediction}
\title{Plot prediction}
\usage{
plot_prediction(obj, data)
}
\arguments{
\item{obj}{model object}

\item{data}{data.frame used to plot and calculate predictions based on}
}
\value{
list with plot and performance stats
}
\description{
Plot prediction
}
