% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.stock}
\alias{plot.stock}
\title{Plotting the trades}
\usage{
\method{plot}{stock}(obj)
}
\arguments{
\item{obj}{object of class stock which is obtained from `get_performance`}
}
\description{
Plotting the trades
}
